requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
chromadb>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit>=1.28.0
konlpy>=0.6.0
soynlp>=1.0.2
nltk>=3.8
spacy>=3.7.0
transformers>=4.35.0
selenium>=4.15.0
scrapy>=2.11.0
Pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
